pyproject.toml
[tool.poetry]
name = "python_magnetgeo"
version = "0.3.0"
description = "Python helpers to create HiFiMagnet cads and meshes"
authors = ["Christophe Trophime <christophe.trophime@lncmi.cnrs.fr>",
           "Romain Vallet <romain.vallet@lncmi.cnrs.fr>",
           "Jeremie Muzet <jeremie.muzet@lncmi.cnrs.fr>"
          ]

[tool.poetry.dependencies]
python = "^3.11"
PyYAML = "5.3.1"
chevron = "^0.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
